Refactos folder structure and config Dockerfile
