+++ transactions/Dockerfile
@@ -1,0 +1,14 @@
+# Utilizar una imagen base de OpenJDK
+FROM openjdk:17-jdk-slim
+
+# Establecer el directorio de trabajo
+WORKDIR /app
+
+# Copiar el archivo JAR de la aplicación al contenedor
+COPY target/devsu-accounts.jar app.jar
+
+# Exponer el puerto en el que se ejecutará la aplicación
+EXPOSE 8081
+
+# Comando para ejecutar la aplicación
+ENTRYPOINT ["java", "-jar", "app.jar"]
