+++ transactions/src/main/java/com/transactions/devsu/model/entities/Account.java
@@ -1,13 +1,10 @@
-package com.transactions.devsu.entities;
+package com.transactions.devsu.model.entities;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,7 +12,6 @@ import lombok.Setter;
 import lombok.ToString;
 import org.hibernate.Hibernate;
 
-import java.time.LocalDateTime;
 import java.util.Objects;
 
 @Entity
@@ -24,34 +20,33 @@ import java.util.Objects;
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public class Movement {
+public class Account {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false)
-    private LocalDateTime date;
+    @Column(unique = true, nullable = false)
+    private Integer accountNumber;
 
     @Column(nullable = false)
-    private String type;
+    private String accountType;
 
     @Column(nullable = false)
-    private Double amount;
+    private Double initialBalance;
 
     @Column(nullable = false)
-    private Double balance;
+    private Boolean status;
 
-    @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "account_id", nullable = false)
-    private Account account;
+    @Column(nullable = false)
+    private Long clientId;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        Movement movement = (Movement) o;
-        return id != null && Objects.equals(id, movement.id);
+        Account account = (Account) o;
+        return id != null && Objects.equals(id, account.id);
     }
 
     @Override
