+++ transactions/src/main/java/com/transactions/devsu/model/entities/Movement.java
@@ -1,10 +1,13 @@
-package com.transactions.devsu.entities;
+package com.transactions.devsu.model.entities;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,6 +15,7 @@ import lombok.Setter;
 import lombok.ToString;
 import org.hibernate.Hibernate;
 
+import java.time.LocalDateTime;
 import java.util.Objects;
 
 @Entity
@@ -20,33 +24,34 @@ import java.util.Objects;
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public class Account {
+public class Movement {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(unique = true, nullable = false)
-    private Integer accountNumber;
+    @Column(nullable = false)
+    private LocalDateTime date;
 
     @Column(nullable = false)
-    private String accountType;
+    private String type;
 
     @Column(nullable = false)
-    private Double initialBalance;
+    private Double amount;
 
     @Column(nullable = false)
-    private Boolean status;
+    private Double balance;
 
-    @Column(nullable = false)
-    private Long clientId;
+    @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "account_id", nullable = false)
+    private Account account;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        Account account = (Account) o;
-        return id != null && Objects.equals(id, account.id);
+        Movement movement = (Movement) o;
+        return id != null && Objects.equals(id, movement.id);
     }
 
     @Override
