+++ transactions/src/main/java/com/transactions/devsu/services/CustomerClient.java
@@ -1,13 +1,13 @@
 package com.transactions.devsu.services;
 
-import com.transactions.devsu.dto.ClientDTO;
+import com.transactions.devsu.model.dto.ClientDTO;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
 @ComponentScan
-@FeignClient(name = "customer-service", url = "http://localhost:8080/clientes")
+@FeignClient(name = "customer-service", url = "${customer.service.url}")
 public interface CustomerClient {
 
     @GetMapping("/{id}")
